move Consumer record to dataplane
